Try to maybe fix opencv?
my understanding is that opencv-headless simply ships binaries compiled without
GL dependencies, so it needs to be installed after installation of non-headless
version which is pulled in by one of the dependencies we have. binary world of
pip is messed up, but compiling this from source is not something I would do
with the current setup.